eclipse/che#11306 move existing API to 'proposed' file
